requirements.txt
requests
urllib3
line-bot-sdk
Flask
pytesseract
Pillow